corriger les erreurs

+++ squelPartie2/frogSquel2/src/environment/Lane.java
@@ -57,8 +57,9 @@ public class Lane {
 
 	/**
 	 * avoir la densite
+	 * @return
 	 */
-	public int getDensity() {
+	public double getDensity() {
 		return density;
 	}
